Wed, Aug 30, 2023, 1:46 PM +02:00

+++ interpreter/interpreter.py
@@ -1,6 +1,6 @@
 x,y,z = input("Expression: ").split(" ")
-x = int(x)
-z = int(z)
+x = float(x)
+z = float(z)
 if y == "+":
     print(x + z)
 elif y == "-":
